server-django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
reportlab==4.0.7
Pillow==10.1.0
bcrypt==4.1.2
